requirements.txt
PyNaCl==1.4.0
ascon==0.4.0
